+++ PlusLib/src/PatternLocAlgo/FidPatternRecognitionCommon.h
@@ -28,33 +28,33 @@ public:
   static bool PositionLessThan( std::vector<Dot>::iterator b1, std::vector<Dot>::iterator b2 );
 
   /*! Set the x coordinate of the dot */
-  void	SetX(float value) { m_X = value; };
+  void	SetX(double value) { m_X = value; };
 
   /*! Get the x coordinate of the dot */
-  float	GetX() { return m_X; };
+  double	GetX() { return m_X; };
 
   /*! Set the y coordinate of the dot */
-  void	SetY(float value) { m_Y = value; };
+  void	SetY(double value) { m_Y = value; };
 
   /*! Get the y coordinate of the dot */
-  float	GetY() { return m_Y; };
+  double	GetY() { return m_Y; };
 
   /*! Set the intensity of the dot */
-  void	SetDotIntensity(float value) { m_DotIntensity = value; };
+  void	SetDotIntensity(double value) { m_DotIntensity = value; };
 
   /*! Get the intensity of the dot */
-  float	GetDotIntensity() { return m_DotIntensity; };
+  double	GetDotIntensity() { return m_DotIntensity; };
 
   /*! Get Euclidean distance from another point */
-  float	GetDistanceFrom(Dot &d);
+  double	GetDistanceFrom(Dot &d);
 
   /*! Compare two dots, coordinate-wise */
   bool operator== (const Dot& data) const { return (m_X == data.m_X && m_Y == data.m_Y) ; }
 
 protected:
-  float	m_X;
-  float	m_Y;
-  float	m_DotIntensity;
+  double	m_X;
+  double	m_Y;
+  double	m_DotIntensity;
 };
 
 //-----------------------------------------------------------------------------
@@ -75,10 +75,10 @@ public:
   static bool compareLines( Line line1, Line line2 );
 
   /*! Compute the angle in radians between the line and the positive x-axis, value between -Pi and +Pi */
-  static float ComputeAngleRad(Line &line1);
+  static double ComputeAngleRad(Line &line1);
 
   /*! Compute the angle difference in radians between two lines, the result value is between 0 and Pi/2 */
-  static float ComputeAngleRad(Line &line1, Line &line2);
+  static double ComputeAngleRad(Line &line1, Line &line2);
 
   /*! Set the point of line
   \param aIndex is the index in the dots vector fo the line
@@ -93,22 +93,22 @@ public:
   std::vector<int>*	GetPoints() { return &m_Points; };
 
   /*! Set the intensity of the line, which is the sum of its dots intensity */
-  void	SetIntensity(float value) { m_Intensity = value; };
+  void	SetIntensity(double value) { m_Intensity = value; };
 
   /*! Get the intensity of the line, which is the sum of its dots intensity */
-  float	GetIntensity() { return m_Intensity; };
+  double	GetIntensity() { return m_Intensity; };
 
   /*! Set the length of the line */
-  void SetLength(float value) { m_Length = value; };
+  void SetLength(double value) { m_Length = value; };
 
   /*! Get the length of the line */
-  float	GetLength() { return m_Length; };
+  double	GetLength() { return m_Length; };
 
   /*! Set the direction vector that defines the line */
-  void	SetDirectionVector(int aIndex, float aValue) { m_DirectionVector[aIndex] = aValue; };
+  void	SetDirectionVector(int aIndex, double aValue) { m_DirectionVector[aIndex] = aValue; };
 
   /*! Get the direction vector that defines the line */
-  const float*	GetDirectionVector() const{ return &(m_DirectionVector[0]); };
+  const double*	GetDirectionVector() const{ return &(m_DirectionVector[0]); };
 
   /*! Set the start point index of the line. It is an index in the m_DotsVector */
   void  SetStartPointIndex(int index) { m_StartPointIndex = index; };
@@ -125,9 +125,9 @@ public:
 
 protected:
   std::vector<int>	m_Points; // indices of points that make up the line
-  float				m_Intensity;
-  float				m_Length;
-  float       m_DirectionVector[2];
+  double				m_Intensity;
+  double				m_Length;
+  double       m_DirectionVector[2];
   int         m_StartPointIndex;//index of startpoint of the line, all the other line points are towards the positive m_DirectionVector direction from this point
   int         m_EndPointIndex;//Index of the endpoint of the line
 
@@ -161,8 +161,8 @@ public:
   virtual ~Pattern() { };
 
   std::vector<Wire> Wires;
-  std::vector<float> DistanceToOriginMm; //These distances are in mm.
-  std::vector<float> DistanceToOriginToleranceMm; //These tolerances are in mm.
+  std::vector<double> DistanceToOriginMm; //These distances are in mm.
+  std::vector<double> DistanceToOriginToleranceMm; //These tolerances are in mm.
 
   //unused iterator to iterate only through one type of Patterns.
   /*
@@ -248,10 +248,10 @@ public:
   std::vector< std::vector<double> >	GetFoundDotsCoordinateValue() { return m_FoundDotsCoordinateValue; };
 
   /*! Set the cumulate intensity of all dots in the pattern */
-  void	SetIntensity(float value) { m_Intensity = value; };
+  void	SetIntensity(double value) { m_Intensity = value; };
 
   /*! Get the cumulate intensity of all dots in the pattern */
-  float	GetIntensity() { return m_Intensity; };
+  double	GetIntensity() { return m_Intensity; };
 
   /*! Set the number of candidate points that have been found */
   void	SetNumDots(double value) { m_NumDots = value; };
@@ -276,7 +276,7 @@ protected:
   values after the morphological operations, with the pixel values on the
   range 0-1.  A good intensity score is over 100. A bad one (but still
   valid) is below 25. */
-  float m_Intensity;
+  double m_Intensity;
 
   /*! number of possibel fiducial points */
   double m_NumDots;
